OCPBUGSM-35978: Add cluster_id and infra_env_id to logs in more flows
- propogate the route params to asyn contexts
- add infra_env_id to logs fields
